Move TreeNodeValid out of core
