+++ packages/dds/tree/src/simple-tree/arrayNode.ts
@@ -33,9 +33,7 @@ import {
 	typeNameSymbol,
 	NodeKind,
 	type TreeNode,
-	TreeNodeValid,
 	type InternalTreeNode,
-	type MostDerivedData,
 	type TreeNodeSchemaClass,
 	type TreeNodeSchema,
 } from "./core/index.js";
@@ -45,6 +43,7 @@ import { getFlexSchema } from "./toFlexSchema.js";
 import { UsageError } from "@fluidframework/telemetry-utils/internal";
 import { assert } from "@fluidframework/core-utils/internal";
 import { getKernel } from "./core/index.js";
+import { TreeNodeValid, type MostDerivedData } from "./treeNodeValid.js";
 
 /**
  * A generic array type, used to defined types like {@link (TreeArrayNode:interface)}.
