+++ packages/dds/tree/src/simple-tree/core/treeNodeKernel.ts
@@ -14,6 +14,7 @@ import {
 	TreeStatus,
 	treeStatusFromAnchorCache,
 } from "../../feature-libraries/index.js";
+import type { NodeKind, TreeNodeSchema } from "./treeNodeSchema.js";
 
 const treeNodeToKernel = new WeakMap<TreeNode, TreeNodeKernel>();
 
@@ -42,6 +43,20 @@ export function isTreeNode(candidate: unknown): candidate is TreeNode | Unhydrat
 }
 
 /**
+ * Returns a schema for a value if the value is a {@link TreeNode}.
+ *
+ * Returns undefined for other values.
+ * @remarks
+ * Does not give schema for a {@link TreeLeafValue}.
+ */
+export function tryGetTreeNodeSchema<T>(
+	value: T,
+): undefined | TreeNodeSchema<string, NodeKind, unknown, T> {
+	const kernel = treeNodeToKernel.get(value as TreeNode);
+	return kernel?.schema as undefined | TreeNodeSchema<string, NodeKind, unknown, T>;
+}
+
+/**
  * Contains state and an internal API for managing {@link TreeNode}s.
  * @remarks All {@link TreeNode}s have an associated kernel object.
  * The kernel has the same lifetime as the node and spans both its unhydrated and hydrated states.
@@ -59,7 +74,10 @@ export class TreeNodeKernel implements Listenable<TreeChangeEvents> {
 	 * @remarks
 	 * Exactly one kernel per TreeNode should be created.
 	 */
-	public constructor(public readonly node: TreeNode) {
+	public constructor(
+		public readonly node: TreeNode,
+		public readonly schema: TreeNodeSchema,
+	) {
 		assert(!treeNodeToKernel.has(node), "only one kernel per node can be made");
 		treeNodeToKernel.set(node, this);
 	}
